pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "metadv"
version = "0.1.0"
description = "Metadata-Driven Data Vault Generator for dbt"
readme = "README.md"
license = "MIT"
requires-python = ">=3.8"
authors = [
    { name = "MetaDV Contributors" }
]
keywords = [
    "dbt",
    "data-vault",
    "datavault",
    "automate-dv",
    "datavault4dbt",
    "sql",
    "generator",
    "data-modeling",
    "data-warehouse"
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Database",
    "Topic :: Software Development :: Code Generators",
]
dependencies = [
    "pyyaml>=5.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "black>=23.0",
    "isort>=5.0",
    "mypy>=1.0",
    "types-PyYAML>=6.0",
]

[project.urls]
Homepage = "https://github.com/data-diving/metadv"
Documentation = "https://github.com/data-diving/metadv#readme"
Repository = "https://github.com/data-diving/metadv"
Issues = "https://github.com/data-diving/metadv/issues"

[project.scripts]
metadv = "metadv.generator:main"

[tool.hatch.build.targets.sdist]
include = [
    "/src",
]

[tool.hatch.build.targets.wheel]
packages = ["src/metadv"]

[tool.black]
line-length = 100
target-version = ["py38", "py39", "py310", "py311", "py312"]

[tool.isort]
profile = "black"
line_length = 100

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
